requirement.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.7
django-bootstrap-v5==1.0.5
django-embed-video==1.4.0
django-render-partial==0.4
idna==3.2
importlib-metadata==3.10.1
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.7
zipp==3.5.0
